Code AP: Win Screen
Added a win screen so that when a player enters a certain area they get sent to the win screen. I placed an empty game object in front of the mausoleum asset for testing purposes only I removed it from that location after doing my video. The endobject is located in the prefabs folder.