Design/Level3Whitebox
I added enemies to level 3, as well as some other items. I also moved some of the obstacle placeholders to allow free exploration for the build. I added deathzones and planes that show the enemies' path for the navmesh. Finally I implemented the UI tutorial in in the gameplay demo scene, a very simple script that makes a text pop out and disappear shortly after.