SebastianG - Leve3/Design
I placed the new assets into the world and replaced most of the whiteboxes with models. I fixed some holes in the cave and did some minor rescaling of some assets to fit the new models better. I started working a bit on lighting in the cave, hallway, and possession room.