MN Art Producer
Setup LVL3 with path lighting, redid torches as the sculpt ground material is all over the place when it comes to displaying light - had to be redone. added lights to the boss areas as well. LVL 3 skybox changed, will still need changing. bloom added to lvl 3 and lvl 2. skyscraper lights in lvl 2 fixed.