SebastianG_Level3/Design
I fixed the terrain on lvl 1 to fit the new area, I also modified it a bit.
On level 3 I made several rock prefabs, some to be used as decorations, one for ranged attack and made some debris for the blocked areas. I also replaced the door placeholders with doors and finished decorating the castle
I replaced the barrels with the rocks