SebastianG_Level1-2-3/Design
Level 1: Removed the tree from the starting area and added dead enemies in the catacombs.
Level 2: Added colliders to the platforming section and added some buildings around to make it look less  empty.
Level 3: Fixed multiple holes and overlaps in the cave, I also fixed the lip on the cave's entrance. Added some assets and modified the terrain a bit.